% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_columns.R
\name{add_tidy_contest}
\alias{add_tidy_contest}
\title{Append column of tidy contest identifiers}
\usage{
add_tidy_contest(df, platform = NULL, sport = NULL)
}
\arguments{
\item{df}{data.frame with contest identifier column}

\item{platform}{character, which platform the contests are from}

\item{sport}{character, which sport contests are from}
}
\value{
data.frame with column "tidy_contest" appended for easier counting
}
\description{
Append column of tidy contest identifiers
}
