% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_columns.R
\name{add_tidy_entry}
\alias{add_tidy_entry}
\title{Append column of tidy contest entry identifiers}
\usage{
add_tidy_entry(df, platform = NULL, sport = NULL)
}
\arguments{
\item{df}{data.frame with entry identifier column}

\item{platform}{character, which platform the entries are from}

\item{sport}{character, which sport entries are from}
}
\value{
data.frame with column "tidy_entry" appended for easier counting
}
\description{
Append column of tidy contest entry identifiers
}
