% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_columns.R
\name{add_tidy_playernames}
\alias{add_tidy_playernames}
\title{Append column of tidy player names}
\usage{
add_tidy_playernames(df, lookup = NULL, platform = NULL,
  sport = NULL)
}
\arguments{
\item{df}{data.frame with player column}

\item{lookup}{character, path to csv lookup}

\item{platform}{character, which dfs platform to use}

\item{sport}{character, which sport}
}
\value{
data.frame with column "tidy_playername" appended for easier joining
}
\description{
Append column of tidy player names
}
