% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_columns.R
\name{add_tidy_projection}
\alias{add_tidy_projection}
\title{Append column of tidy projections}
\usage{
add_tidy_projection(df, column = NULL)
}
\arguments{
\item{df}{data.frame with position column}

\item{column}{character, which column to use}
}
\value{
data.frame with column "tidy_projection" appended for easier joining
}
\description{
Append column of tidy projections
}
