% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_columns.R
\name{add_tidy_teamabbrev}
\alias{add_tidy_teamabbrev}
\title{Append column of tidy team abbreviations}
\usage{
add_tidy_teamabbrev(
  df = NULL,
  lookup = system.file("extdata", "teamabbrev_lu.csv", package = "dfstools"),
  platform = NULL,
  sport = NULL
)
}
\arguments{
\item{df}{data.frame with team abbreviation column}

\item{lookup}{character, path to csv lookup}

\item{platform}{character, which dfs platform to use}

\item{sport}{character, which sport}
}
\value{
data.frame with column "tidy_teamabbrev" appended for easier joining
}
\description{
Append column of tidy team abbreviations
}
