% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{archetypal_analysis}
\alias{archetypal_analysis}
\title{Archetypal Analysis}
\usage{
archetypal_analysis(pbs)
}
\arguments{
\item{pbs}{player box score}
}
\value{
a list of
\itemize{
\item archetypes the parameters that define each archetype
\item players the players tagged with their loadings on each archetype
\item best the model object created in the analysis
\item archdata the numerical inputs for the model}
}
\description{
perform an "archetypal athletes" analysis
}
