% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{archetype_prep}
\alias{archetype_prep}
\title{Archetype Prep}
\usage{
archetype_prep(pbs)
}
\arguments{
\item{pbs}{player box score}
}
\value{
an "as" object from archetypes::stepArchetypes
}
\description{
prepare a plaer box score for archetypal analysis
}
