% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypoid_models.R
\name{archetypoid_search}
\alias{archetypoid_search}
\title{Archetypoid Search}
\usage{
archetypoid_search(player_totals, player_labels, num_steps = 1:15,
  nrep = 8, verbose = FALSE)
}
\arguments{
\item{player_totals}{a tibble of the values to use for archetypal analysis.
The first column must be the player name, which must match the player name
in the player_labels. The archetypoids will be sorted on the second columns.}

\item{player_labels}{a tibble with the labels (player and team names,
positions, etc.). They may overlap with the totals. The leftmost column must
be the player name, which must match the player name in \code{player_totals}..}

\item{num_steps}{number of steps to use (default 1:5)}

\item{nrep}{number of repetitions at each step (default 4)}

\item{verbose}{should the search be verbose? (default FALSE)}
}
\value{
a list of
\itemize{
\item archetypoid_parameters the parameters that define each archetypoid
\item player_alphas the players tagged with their loadings on each archetypoid
\item archetypoid_model the model object - the \code{bestModel} with \code{num_steps}
archetypoids
\item all of the models}
}
\description{
stepwise search of archetypoid counts
}
