% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{bdb_wnba_dfs}
\alias{bdb_wnba_dfs}
\title{read BigDataBall WNBA DFS spreadsheet}
\usage{
bdb_wnba_dfs(excel_file)
}
\arguments{
\item{excel_file}{a BigDataBall Excel DFS file}
}
\value{
a tibble with the DFS data
}
\description{
Reads a BigDataBall WNBA DFS spreadsheet to a tibble
}
