% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{bdb_wnba_rank_model}
\alias{bdb_wnba_rank_model}
\title{BigDataBall WNBA mvglmmRank model}
\usage{
bdb_wnba_rank_model(season_team_feed)
}
\arguments{
\item{season_team_feed}{a BigDataBall season team feed tibble}
}
\value{
the model returned by dfstools::mvglmmRank_model
}
\description{
creates a WNBA mvglmmRank model
}
