% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{bdb_wnba_schedule}
\alias{bdb_wnba_schedule}
\title{read BigDataBall WNBA schedule spreadsheet}
\usage{
bdb_wnba_schedule(excel_file)
}
\arguments{
\item{excel_file}{a BigDataBall Excel schedule file}
}
\value{
a tibble with the schedule
}
\description{
Reads a BigDataBall schedule spreadsheet to a tibble
}
