% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{bdb_wnba_team}
\alias{bdb_wnba_team}
\title{read BigDataBall WNBA team feed spreadsheet}
\usage{
bdb_wnba_team(excel_file)
}
\arguments{
\item{excel_file}{a BigDataBall season team Excel file}
}
\value{
a tibble with the team data
}
\description{
Reads a BigDataBall WNBA DFS spreadsheet to a tibble
}
