% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{build_model}
\alias{build_model}
\title{Build model}
\usage{
build_model(season, method = "PB1", first.order = TRUE, OT.flag = TRUE)
}
\arguments{
\item{season}{a tibble returned from the Stattleship API via `get_games`}

\item{method}{method to be passed to mvglmmRank - default is "PB1"}

\item{first.order}{flag to be passed to mvglmmRank - default is TRUE}

\item{OT.flag}{flag to be passes to mvglmmRank - default is TRUE}
}
\value{
an mvglmmRank model object
}
\description{
runs mvglmmRank::mvglmmRank against a game box score and points function
}
