% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conference_table}
\alias{conference_table}
\title{WNBA 2020 table of conferences}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{conference}{conference name}
\item{team}{team name}
\item{code}{team code name}
}
}
\source{
\url{https://www.basketball-reference.com/wnba/}
}
\usage{
conference_table
}
\description{
A dataset with the list of conferences / teams
}
\keyword{datasets}
