% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{connect_database_file}
\alias{connect_database_file}
\title{Connect Database File}
\usage{
connect_database_file(path)
}
\arguments{
\item{path}{full path to file. It will be created if it doesn't exist}
}
\value{
a DBI connection object
}
\description{
creates an SQLite database file if needed and returns its connection
}
\examples{
\dontrun{
library(dfstools)
connection <- connect_database_file("~/DFS/nba/dfs_database.sqlite")
}
}
