% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{create_database}
\alias{create_database}
\title{Create Database}
\usage{
create_database(directory, league, season, verbose = TRUE)
}
\arguments{
\item{directory}{valid path to a directory where the database file
will be stored. It will be created if it doesn't exist.}

\item{league}{the league ("nba", "nhl", "nfl" or "mlb")}

\item{season}{a valid season descriptor for the league}

\item{verbose}{print status information}
}
\value{
a DBI connection object pointing to the database.
\code{create_database} will disconnect before returning.
}
\description{
Creates an SQLite database for a league and season
}
\details{
The database will be saved to file
\code{<directory>/<league>_<season>.sqlite}"`. It will be overwritten if
it exists and created if it doesn't.
}
