% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest_history.R
\name{create_draftkings_database}
\alias{create_draftkings_database}
\title{Create DraftKings Database}
\usage{
create_draftkings_database(draftkings_contest_history, database_file)
}
\arguments{
\item{draftkings_contest_history}{DraftKings contest entry history tibble}

\item{database_file}{a file for the SQLite database}
}
\value{
a connection object to the database
}
\description{
Create a new database from a contest history tibble
}
\examples{
\dontrun{
dk_history <- dktidy::read_draftkings_contest_history(
"~/Projects/draftkings/contest-standings/draftkings-contest-entry-history.csv")
db_connection <- create_draftkings_database(dk_history, "dk_database.sqlite3")
}
}
