% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_utils.R
\name{create_nba_database}
\alias{create_nba_database}
\title{Create NBA Database}
\usage{
create_nba_database(sqlite_file)
}
\arguments{
\item{sqlite_file}{a valid file path; it will be overwritten if it exists and created if it doesn't}
}
\value{
a DBI connection object pointing to the database
}
\description{
Creates an SQLite database for all completed NBA seasons with DFS data
}
