% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_optimized_tiers_pm.R
\name{create_optimized_tiers_pm}
\alias{create_optimized_tiers_pm}
\title{create optimized lineups for draftkings tiers}
\usage{
create_optimized_tiers_pm(
  to_optimize = NULL,
  lineups = NULL,
  max_exposure = NULL,
  max_repeated_players = NULL,
  platform = NULL,
  sport = NULL
)
}
\arguments{
\item{to_optimize}{data.frame of player projections, salary and roster info}

\item{lineups}{number of lineups to create}

\item{max_exposure}{number maximum exposure %}

\item{max_repeated_players}{number NOT CURRENTLY USED}

\item{platform}{character should always be "draftkings-tiers-pm"}

\item{sport}{character}
}
\value{
data.frame
}
\description{
create optimized lineups for draftkings tiers
}
