% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings_standings_utils.R
\name{cumulant_tables}
\alias{cumulant_tables}
\title{Make cumulant tables}
\usage{
cumulant_tables(pbs, pfunc, order_max = 4)
}
\arguments{
\item{pbs}{a player box score}

\item{pfunc}{a points function}

\item{order_max}{maximum cumulant order (default 4)}
}
\value{
a list of
\itemize{
\item minutes_cumulants the cumulants of player minutes, with row names identifying the players
\item fpoints_cumulants the cumulants of player fantasy points, with row names identifying the players}
}
\description{
compute the cumulants of all the players
}
