% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{draftkings_points}
\alias{draftkings_points}
\title{DraftKings fantasy points}
\usage{
draftkings_points(boxscore)
}
\arguments{
\item{boxscore}{a player or game box score data frame}
}
\value{
a vector of DraftKings fantasy points
}
\description{
Computes DraftKings fantasy points from a box score
}
\examples{
\dontrun{
playerbox <- playerboxscore(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
dk_points <- draftkings_points(playerbox)
}
}
