% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{fanduel_points}
\alias{fanduel_points}
\title{FanDuel fantasy points}
\usage{
fanduel_points(boxscore)
}
\arguments{
\item{boxscore}{a player or game box score data frame}
}
\value{
a vector of FanDuel fantasy points
}
\description{
Computes FanDuel fantasy points from a box score
}
\examples{
\dontrun{
playerbox <- playerboxscore(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
fd_points <- fanduel_points(playerbox)
}
}
