% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{game_box_score}
\alias{game_box_score}
\title{Make game box score data frame from player box score}
\usage{
game_box_score(pbs)
}
\arguments{
\item{pbs}{a player box score data frame}
}
\value{
a data frame of game box scores
}
\description{
Creates a game box score data frame from a player box score data frame
}
\examples{
\dontrun{
pbs2015 <- player_box_score(
  "~/Dropbox/15-16-player-data/season-player-feed-01-28-2016.xlsx")
gbs2015 <- game_box_score(pbs2015)
}
}
