% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{game_data}
\alias{game_data}
\alias{collect_game_data}
\title{Collect game data}
\usage{
collect_game_data(gbs, pmg, end_date)
}
\arguments{
\item{gbs}{game box score data frame}

\item{pmg}{player minutes per game (240 for NBA, 200 for WNBA)}

\item{end_date}{the latest date to use as input}
}
\value{
a game data dataframe
}
\description{
creates an input game data file for mvglmmRank
}
