% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{game_score}
\alias{game_score}
\title{Game Score (GmSc)}
\usage{
game_score(boxscore)
}
\arguments{
\item{boxscore}{a player or game box score data frame}
}
\value{
a vector of game scores
}
\description{
Computesthe "game score" metric from a box score
}
\examples{
\dontrun{
playerbox <- player_box_score(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
gmsc <- game_score(playerbox)
}
}
