% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{gameboxscore}
\alias{gameboxscore}
\title{Make game box score data frame from player box score}
\usage{
gameboxscore(pbs)
}
\arguments{
\item{pbs}{a player box score data frame}
}
\value{
a data frame of game box scores
}
\description{
Creates a game box score data frame from a player box score data frame
}
\examples{
\dontrun{
pbs2015 <- playerboxscore(
  "~/Dropbox/15-16-player-data/season-player-feed-01-28-2016.xlsx")
gbs2015 <- gameboxscore(pbs2015)
}
}
