% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fantasypros_projections.R
\name{get_fantasypros_projections}
\alias{get_fantasypros_projections}
\title{Get dfs player projections from fantasypros.com}
\usage{
get_fantasypros_projections(
  sport = NULL,
  from_csv = NULL,
  supported_sports = c("nfl"),
  scraped_headers = c("rank", "wsis", "name", "opp", "best", "worst", "avg", "stdev",
    "proj", "notes"),
  csv_headers = c("RK", "PLAYER NAME", "TEAM", "OPP", "MATCHUP (?)", "START/SIT",
    "PROJ. FPTS"),
  nfl_positions = c("qb", "rb", "wr", "te", "dst")
)
}
\arguments{
\item{sport}{character, which sport to use? supports nfl so far, with nba on the way}

\item{from_csv}{character, if not NULL then path to the csv or csv's with projections}

\item{supported_sports}{character vector of leagues supported}

\item{csv_headers}{character, names of the columns in rotogrinders csv's}

\item{nfl_positions}{character, position names}
}
\value{
data.frame of players and projections for selected platform
}
\description{
Get dfs player projections from fantasypros.com
}
