% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_game_logs}
\alias{get_game_logs}
\title{Get Current Season Game Logs from Stattleship API}
\usage{
get_game_logs(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a tibble with the game_logs for the season.
}
\description{
Gets the game_logs for a league's current season from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_game_logs <-
  tidysportsfeeds::get_game_logs(league = "nba")
nhl_game_logs <-
  tidysportsfeeds::get_game_logs(league = "nhl")
nfl_game_logs <-
  tidysportsfeeds::get_game_logs(league = "nfl")
}
}
