% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_games}
\alias{get_games}
\title{Get current season games from Stattleship API}
\usage{
get_games(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a `games table from the Stattleship API. The entire season is
returned, with column `status` designating "closed", "upcoming" and
"in_progress" games.
}
\description{
Gets a `games` table from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(dfstools)
library(stattleshipR)
stattleshipR::set_token(token)
nba_games <-
  get_games(league = "nba")
nhl_games <-
  get_games(league = "nhl")
nfl_games <-
  get_games(league = "nfl")
}
}
