% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_nba_game_logs}
\alias{get_nba_game_logs}
\title{Get NBA game logs to date, tidied}
\usage{
get_nba_game_logs(nba_data)
}
\arguments{
\item{nba_data}{a list returned by get_game_logs_to_date for the NBA}
}
\value{
a tidied player game log tibble (player box scores)
}
\description{
Gets a tidy game log tibble for the NBA season to date
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
stattleshipR::set_token(token)
nba_data <- get_game_logs_to_date(league = "nba")
nba_game_logs <-
get_nba_game_logs(nba_data)
}
}
