% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_pbp_log}
\alias{get_pbp_log}
\title{Get Current Season Play-by-Play Log from Stattleship API}
\usage{
get_pbp_log(league, game_slug)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}

\item{game_slug}{game slug from the "games" data frame for the season}
}
\value{
a tibble with the play-by-play log for the game
}
\description{
Gets the play-by-play log for a current season game from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nhl_pbp_log <- tidysportsfeeds::get_pbp_log(
  league = "nhl", game_slug = "nhl-2017-2018-phi-sj-2017-10-4-1930")
}
}
