% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_player_table}
\alias{get_player_table}
\title{Get player table from Stattleship API}
\usage{
get_player_table(players_object)
}
\arguments{
\item{players_object}{a "players" table returned by the Stattleship API}
}
\value{
a simplified table suitable for dplyr "join" operations
}
\description{
Gets a joinable player table from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
stattleshipR::set_token(token)
nba_data <- tidysportsfeeds::get_game_logs_to_date(league = "nba")
nba_players <-
tidysportsfeeds::get_player_table(nba_data$season_tables$players)
nba_game_logs <- nba_data$raw_game_logs \%>\%
dplyr::left_join(nba_data$season_tables$games, by = c("game_id" = "id")) \%>\%
dplyr::left_join(nba_players, by = c("player_id" = "id")) \%>\%
dplyr::select(
player = name.y,
birth_date:years_of_experience,
scoreline, started_at, ended_at, period,
is_home_team,
assists:triple_double)
}
}
