% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_players}
\alias{get_players}
\title{Get Current Season PLayers from Stattleship API}
\usage{
get_players(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a tibble with the players for the season.
}
\description{
Gets the players for a league's current season from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_players <-
  tidysportsfeeds::get_players(league = "nba")
nhl_players <-
  tidysportsfeeds::get_players(league = "nhl")
nfl_players <-
  tidysportsfeeds::get_players(league = "nfl")
}
}
