% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rotogrinders_projections.R
\name{get_rotogrinders_projections}
\alias{get_rotogrinders_projections}
\title{Get dfs player projections from Rotogrinders.com}
\usage{
get_rotogrinders_projections(sport = NULL, platform = NULL)
}
\arguments{
\item{sport}{character, which sport to use? supports nfl so far, with nba on the way}

\item{platform}{character, which dfs platform to use? supports draftkings so far, with fanduel on the way}
}
\value{
data.frame of players and projections for selected platform
}
\description{
Get dfs player projections from Rotogrinders.com
}
