% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysportsfeeds_utils.R
\name{get_schedule_gameentry}
\alias{get_schedule_gameentry}
\title{Get Schedule Gameentry}
\usage{
get_schedule_gameentry(league = "nba", season = "2017-2018-regular")
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}

\item{season}{e.g., "2017-2018-regular"}
}
\value{
the gameentry object
}
\description{
Gets a schedule 'gameentry' object from the MySportsFeeds.com API
}
\examples{
\dontrun{
nba_gameentry <-
tidysportsfeeds::get_schedule_gameentry(league = "nba", season = "2017-2018-regular")
nhl_gameentry <-
tidysportsfeeds::get_schedule_gameentry(league = "nhl", season = "2017-2018-regular")
nfl_gameentry <-
tidysportsfeeds::get_schedule_gameentry(league = "nfl", season = "2017-regular")
}
}
