% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_season}
\alias{get_season}
\title{Get Current Season Games from Stattleship API}
\usage{
get_season(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a tibble with the games for the season. The whole schedule is given,
with games as yet unplayed having scores of zero.
}
\description{
Gets a schedule 'gameentry' object from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_games <-
  tidysportsfeeds::get_season(league = "nba")
nhl_games <-
  tidysportsfeeds::get_season(league = "nhl")
nfl_games <-
  tidysportsfeeds::get_season(league = "nfl")
}
}
