% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_teams}
\alias{get_teams}
\title{Get Current Season Teams from Stattleship API}
\usage{
get_teams(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a tibble with the teams for the season.
}
\description{
Gets the teams for a league's current season from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_teams <-
  tidysportsfeeds::get_teams(league = "nba")
nhl_teams <-
  tidysportsfeeds::get_teams(league = "nhl")
nfl_teams <-
  tidysportsfeeds::get_teams(league = "nfl")
}
}
