% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{install_api_wrappers}
\alias{install_api_wrappers}
\title{Install MySportsFeeds and Stattleship Wrappers}
\usage{
install_api_wrappers()
}
\value{
list of return values from the wrapper installs
}
\description{
Installs the MySportsFeed and Stattleship R wrappers from GitHub
}
\examples{
\dontrun{
install_status <- tidysportsfeeds::install_api_wrappers()
}
}
