% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mysportsfeeds_wrapper.R
\name{install_mysportsfeeds_wrapper}
\alias{install_mysportsfeeds_wrapper}
\title{Install MySportsFeeds Wrapper}
\usage{
install_mysportsfeeds_wrapper()
}
\value{
a vector of points
}
\description{
Installs the MySportsFeed R wrapper from GitHub
}
\examples{
\dontrun{
tidysportsfeeds::install_mysportsfeeds_wrapper()
}
}
