% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{kaggle_game_data}
\alias{kaggle_game_data}
\title{Create a \code{game.data} tibble from Kaggle NCAA data}
\usage{
kaggle_game_data(compact_results)
}
\arguments{
\item{compact_results}{a Kaggle NCAA "Compact Results" dataset}
}
\value{
the dataset augmented with the input columns for mvglmmRank
}
\description{
\code{mvglmmRank} reads an input file in a specific \code{game.data}
format. This function takes a "compact results" dataset for an NCAA season
and adds columns to satisfy the \code{mvglmmRank} requirements.
}
\examples{
\dontrun{
RegularSeasonCompactResults <- readr::read_csv(
  "~/DFS/kaggle/RegularSeasonCompactResults.csv",
  col_types = cols(
    DayNum = col_integer(),
    LScore = col_integer(),
    LTeamID = col_integer(),
    NumOT = col_integer(),
    Season = col_integer(),
    WScore = col_integer(),
    WTeamID = col_integer())) \%>\%
  filter(Season == 2019)
}
}
