% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{kaggle_probability}
\alias{kaggle_probability}
\title{Get probability for a Kaggle "game_id"}
\usage{
kaggle_probability(model, game_id)
}
\arguments{
\item{model}{a model returned by \code{mvglmmRank_model}}

\item{game_id}{a Kaggle "game_id" - season, team 1 ID and team 2 ID
separated by underscores}
}
\value{
the probability of a win for team 1
}
\description{
Parses the probability from a \code{game.pred} result
}
