% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{kaggle_ratings}
\alias{kaggle_ratings}
\title{Kaggle ratings}
\usage{
kaggle_ratings(model, teams)
}
\arguments{
\item{model}{an \code{mvglmmRank} model object with a binary ratings object}

\item{teams}{a table of team IDs and names}
}
\value{
a ratings tibble
}
\description{
Compute a team ratings tibble from an mvglmmRank model. The
model must have a \code{b.ratings} object - the method used must have "B" in
it.
}
