% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kickoff_optimization.R
\name{kickoff_optimization}
\alias{kickoff_optimization}
\title{Kickoff dfs optimization pipeline}
\usage{
kickoff_optimization(
  SPORT = NULL,
  PLATFORM = NULL,
  PROJECTIONS = NULL,
  PROJ_COL = NULL,
  LU_MULT = NULL,
  PROJ_PATH = NULL,
  SALARY_PATH = NULL,
  ENTRY_PATH = NULL,
  TEMP_PATH = NULL,
  MAX_EXP = NULL,
  MAX_REPEATED_PLAYERS = NULL
)
}
\arguments{
\item{SPORT}{character, which dfs sport/league}

\item{PLATFORM}{character, which dfs platform}

\item{PROJECTIONS}{character, which projections to use}

\item{PROJ_COL}{character, which column in projections}

\item{LU_MULT}{numeric, LU_MULT * number of entries = number of lineups generated}

\item{PROJ_PATH}{character, path to projections csv}

\item{SALARY_PATH}{character, path to player salary csv}

\item{ENTRY_PATH}{character, path to entry csv}

\item{TEMP_PATH}{character, path to directory for writing temp files}

\item{MAX_EXP}{numeric, maximum exposure to any individual player in lineups}

\item{MAX_REPEATED_PLAYERS}{numeric, maximum number of repeated combinations of players}
}
\value{
outputs a csv to be entered into dfs site
}
\description{
Kickoff dfs optimization pipeline
}
