% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings_standings_utils.R
\name{lineup_moments}
\alias{lineup_moments}
\title{Lineup moments calculation}
\usage{
lineup_moments(standings, pbs, before)
}
\arguments{
\item{standings}{a data frame from a DraftKings standings CSV file}

\item{pbs}{a player box score}

\item{before}{date (character, 'yyyy-mm-dd') - only use data before this date}
}
\value{
returns stuff
}
\description{
compute the mean and variance of a lineup
}
