% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{make_game_data}
\alias{make_game_data}
\title{Make game.data}
\usage{
make_game_data(seasonal_games)
}
\arguments{
\item{seasonal_games}{a tibble returned from \code{msf_seasonal_games}}
}
\value{
a "game data" table
}
\description{
Builds a \code{mvglmmRank} "game.data" table from an
\code{msf_seasonal_games}result
}
\examples{
\dontrun{
nhl_seasonal_games <- dfstools::msf_seasonal_games("nhl", "2019-2020-regular")
nhl_game_data <- dfstools::make_game_data(nhl_seasonal_games)
nhl_model <- dfstools::mvglmmRank_model(
  nhl_game_data, method = "PB1", first.order = FALSE, verbose = TRUE
)
nhl_schedule <- dfstools::make_schedule(nhl_seasonal_games)
nhl_predictions <- dfstools::game_predict(nhl_schedule, nhl_model)
View(nhl_predictions)
entropy_forecast <- nhl_predictions \%>\%
  group_by(date) \%>\%
  summarize(games = n(), entropy = sum(entropy))
View(entropy_forecast)
}
}
