% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{make_schedule}
\alias{make_schedule}
\title{Make schedule}
\usage{
make_schedule(seasonal_games)
}
\arguments{
\item{seasonal_games}{a tibble returned from \code{msf_seasonal_games}}
}
\value{
a "schedule" table
}
\description{
Builds a \code{game_predict} "schedule" table from an
\code{msf_seasonal_games}result
}
\examples{
\dontrun{
nba_seasonal_games <- dfstools::msf_seasonal_games("nba", "2018-2019-regular")
nhl_seasonal_games <- dfstools::msf_seasonal_games("nhl", "2019-2020-regular")
nba.schedule <- dfstools::make_schedule(nba_seasonal_games)
nhl.schedule <- dfstools::make_schedule(nhl_seasonal_games)
}
}
