% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_build_url}
\alias{msf_build_url}
\title{Build MySportsFeeds API URL}
\usage{
msf_build_url(league, season, endpoint, team = "", date = "")
}
\arguments{
\item{league}{c("nba", "nhl", "nfl", "mlb")}

\item{season}{string the season}

\item{endpoint}{string the JSON endpoint, for example "games.json"}

\item{team}{string for feeds that must be subsetted by team (default "")}

\item{date}{string for feeds that must be subsetted by date (default "")}
}
\value{
the MySportsFeed URL
}
\description{
builds a MySportsFeeds 2.1 API URL
}
\examples{
\dontrun{
nba_games_url <- dfstools::msf_build_url(
  "nba",
  "2018-2018-regular",
  "games.json"
)
redwings_gamelogs <- dfstools::msf_build_url(
  "nhl",
  "2018-2019-regular",
  "team_gamelogs.json",
  team = "DET"
)
xmas_nba_games <- msf_build_url(
  "nba",
  "2019-2020-regular",
  "games.json",
  date = "20191225"
)
}
}
