% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_daily_dfs}
\alias{msf_daily_dfs}
\title{MySportsFeeds Daily DFS}
\usage{
msf_daily_dfs(league, season, date, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{date}{the date to fetch}

\item{verbose}{print status info}
}
\value{
a \code{dfs} object
}
\description{
Gets DFS object from from
MySportsFeeds version 2.1 API
}
\examples{
\dontrun{
nba_dfs <- dfstools::msf_daily_dfs(
  season = "2019-playoff", league = "nba", date = "20190419", verbose = TRUE
)}
}
