% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_get_feed}
\alias{msf_get_feed}
\title{GET from MySportsFeeds API}
\usage{
msf_get_feed(url, verbose)
}
\arguments{
\item{url}{the URL to GET}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success)
\item response if status_code == 200, the JSON object;
otherwise, the raw text.
}
}
\description{
GETs data from the MySportsFeeds 2.1 API feed
}
\examples{
\dontrun{
response <- dfstools::msf_get_feed(
"https://api.mysportsfeeds.com/v2.1/pull/nba/2018-playoff/games.json"
)}
}
