% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_nba_dfs}
\alias{msf_nba_dfs}
\title{MySportsFeeds NBA DFS}
\usage{
msf_nba_dfs(season, date, apikey)
}
\arguments{
\item{season}{the season to fetch}

\item{date}{the date to fetch}

\item{apikey}{your MySportsFeeds API key (version 2.0!)}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success, -200 for success but no DFS data))
\item response if status_code == 200, a tibble of DFS data; otherwise, the raw text.
}
}
\description{
Gets DFS data object from from MySportsFeeds version 2.0 API
}
\examples{
\dontrun{
apikey <- "your_API key"
library(dfstools)
nba_dfs <- msf_nba_dfs("2018-playoff", "20180414", apikey)
}
}
