% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_nba_games}
\alias{msf_nba_games}
\title{MySportsFeeds NBA Games}
\usage{
msf_nba_games(season, apikey)
}
\arguments{
\item{season}{the season to fetch games for}

\item{apikey}{your MySportsFeeds API key (version 2.0!)}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success))
\item response if status_code == 200, a tibble of games; otherwise, the raw text.
}
}
\description{
Returns a data frame of NBA games from MySportsFeeds version 2.0 API
}
\details{
`msf_nba_games` adds two columns at the right of the tibble:
  `season`, the source season of the data, and `date`, the game
  date (started) in the Eastern USA timezone ("EST5EDT"). The returned
  tibble will be sorted in chronological order.
}
\examples{
\dontrun{
apikey <- "your_API key"
library(dfstools)
nba_games <- msf_nba_games(season = "2017-2018-regular", apikey)
}
}
