% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_past_nba_dfs}
\alias{msf_past_nba_dfs}
\title{MySportsFeeds Past NBA DFS}
\usage{
msf_past_nba_dfs(nba_games)
}
\arguments{
\item{nba_games}{a tibble produced with `msf_past_nba_games`}
}
\value{
a tibble with all the DFS data available for the input games
}
\description{
Gets DFS data object from the MySportsFeeds.com API
}
\examples{
\dontrun{
apikey <- "your_API key"
library(dfstools)
library(mysportsfeedsR)
authenticate_v2_x(apikey)
seasons <- c(
  "2018-playoff",
  "2017-2018-regular",
  "2017-playoff",
  "2016-2017-regular",
  "2016-playoff",
  "2015-2016-regular")
nba_games <- msf_past_nba_games(seasons)
nba_dfs <- msf_past_nba_dfs(nba_games)
}
}
