% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_retry_call}
\alias{msf_retry_call}
\title{MySportsFeeds Retry Call}
\usage{
msf_retry_call(ntries, sleep_seconds, function_name, league, season, team,
  apikey, verbose)
}
\arguments{
\item{ntries}{number of times to try the call}

\item{sleep_seconds}{seconds to sleep after failure}

\item{function_name}{the function to call,}

\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{team}{the team to fetch}

\item{apikey}{your MySportsFeeds API key (version 2.0!)}

\item{verbose}{print status if TRUE}
}
\value{
the last response from the call, successful or not
}
\description{
Retries calls till they work or a specified number of failures
}
