% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_player_gamelogs}
\alias{msf_seasonal_player_gamelogs}
\title{MySportsFeeds Seasonal Player Gamelogs}
\usage{
msf_seasonal_player_gamelogs(league, season, team, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{team}{the team to fetch}

\item{verbose}{print status info}
}
\value{
a \code{player_gamelogs} object
}
\description{
Gets player gamelogs object from from
MySportsFeeds version 2.1 API
}
\examples{
\dontrun{
nba_player_gamelogs <- dfstools::msf_seasonal_player_gamelogs(
  season = "2018-playoff", league = "nba", team = "GSW"
)}
}
