% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_player_stats_totals}
\alias{msf_seasonal_player_stats_totals}
\title{MySportsFeeds Seasonal Player Stats Totals}
\usage{
msf_seasonal_player_stats_totals(league, season, verbose)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success,
-200 for HTTP success but no data)
\item  response if status_code == 200, a \code{player_stats_totals}
object; otherwise, the raw text.
}
}
\description{
Gets player gamelogs object from from
MySportsFeeds version 2.1 API
}
\examples{
\dontrun{
nba_player_totals <- dfstools::msf_seasonal_player_stats_totals(
season = "current", league = "nba", verbose = TRUE
)}
}
