% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_players}
\alias{msf_seasonal_players}
\title{MySportsFeeds Seasonal Players}
\usage{
msf_seasonal_players(league, season, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{verbose}{print status info}
}
\value{
a tibble of players
}
\description{
Returns a data frame of players from
MySportsFeeds version 2.1 API
}
\examples{
\dontrun{
nba_players <- dfstools::msf_seasonal_players(
season = "2017-2018-regular", league = "nba"
)}
}
