% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_team_gamelogs}
\alias{msf_seasonal_team_gamelogs}
\title{MySportsFeeds Seasonal Team Gamelogs}
\usage{
msf_seasonal_team_gamelogs(league, season, team, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{team}{the team to fetch}

\item{verbose}{print status info}
}
\value{
a \code{team_gamelogs} object
}
\description{
Gets team gamelogs object from from MySportsFeeds
version 2.1 API
}
\examples{
\dontrun{
nba_team_gamelogs <- dfstools::msf_seasonal_team_gamelogs(
season = "2018-playoff", league = "nba", team = "GSW"
)}
}
