% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasons}
\alias{msf_seasons}
\title{MySportsFeeds Seasons}
\usage{
msf_seasons()
}
\value{
a tibble where each row is a league / season pair
}
\description{
list the league / season combinations for which
MySportsFeeds has DFS data
}
\examples{
\dontrun{
seasons <- dfstools::msf_seasons()
seasons
}
}
