% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{mvglmmRank_model}
\alias{mvglmmRank_model}
\title{Build an mvglmmRank model}
\usage{
mvglmmRank_model(
  game_data,
  method = "NB",
  first.order = FALSE,
  verbose = FALSE,
  OT.flag = FALSE
)
}
\arguments{
\item{game_data}{a \code{game_data} tibble.}

\item{method}{the \code{mvglmmRank} method to use - default is "NB"}

\item{first.order}{the \code{mvglmmRank} first-order correction flag -
default is \code{FALSE}}

\item{verbose}{print a lot of stuff whilst iterating - default is \code{FALSE}}

\item{OT.flag}{the \code{mvglmmRank} "use overtime" flag - default is \code{FALSE}}
}
\value{
an mvglmmRank model object
}
\description{
Builds an mvglmmRank model
}
\examples{
\dontrun{
nhl_seasonal_games <- dfstools::msf_seasonal_games("nhl", "2019-2020-regular")
nhl_game_data <- dfstools::make_game_data(nhl_seasonal_games)
nhl_model <- dfstools::mvglmmRank_model(
  nhl_game_data, method = "PB1", first.order = FALSE, verbose = TRUE
)
nhl_schedule <- dfstools::make_schedule(nhl_seasonal_games)
nhl_predictions <- dfstools::game_predict(nhl_schedule, nhl_model)
View(nhl_predictions)
entropy_forecast <- nhl_predictions \%>\%
  group_by(date) \%>\%
  summarize(games = n(), entropy = sum(entropy))
View(entropy_forecast)
}
}
