% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{nba_archetype_search}
\alias{nba_archetype_search}
\title{NBA Archetype Search}
\usage{
nba_archetype_search(player_totals, num_steps = 7)
}
\arguments{
\item{player_totals}{a tibble returned by \code{nba_player_season_totals}}

\item{num_steps}{number of steps to use (default 7)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object - the \code{bestModel} with \code{num_steps}
archetypes
\item all of the models}
}
\description{
stepwise search of archetype counts
}
\examples{
\dontrun{
dfstools::msf_set_apikey("your MySportsFeeds API key")
player_totals <- dfstools::nba_player_season_totals("current")
the_archetypes <- dfstools::nba_archetype_search(player_totals)
player_alphas <- the_archetypes[["player_alphas"]]
View(player_alphas)
}
}
