% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_analytics.R
\name{nba_play_by_play}
\alias{nba_play_by_play}
\title{NBA Play-by-play}
\usage{
nba_play_by_play(season, game_slug)
}
\arguments{
\item{season}{the NBA season}

\item{game_slug}{a string of the form \code{yyyymmdd-AAA-HHH}, where \code{yyyymmdd}
is the game start date, \code{AAA} is the away team abbreviation, and \code{HHH} is the
home team abbreviation.}
}
\value{
a tibble of all the plays in the game
}
\description{
retrieve an nba game play-by-play tibble
}
\examples{
\dontrun{
library(dplyr)
nba_playoff_games <- dfstools::msf_seasonal_games("nba", "2019-playoff")
nba_playoff_final <- nba_playoff_games \%>\%
  dplyr::arrange(desc(slug)) \%>\%
  head(1)
nba_playoff_final_slug <- nba_playoff_final[["slug"]]
nba_playoff_final_pbp <-
  dfstools::nba_play_by_play("2019-playoff", nba_playoff_final_slug)
}
}
