% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_analytics.R
\name{nba_player_season_totals}
\alias{nba_player_season_totals}
\title{NBA Player Season Totals}
\usage{
nba_player_season_totals(season)
}
\arguments{
\item{season}{a valid MySportsFeeds v2.1 API season name}
}
\value{
a list of two items
\itemize{
\item player_totals a tibble of player season total box score statistics,
arranged by descending points scored
\item player_labels a tibble of labeling information for players
}
}
\description{
fetches the player season totals for base box score statistics
}
\examples{
\dontrun{
player_totals <- dfstools::nba_player_season_totals("current")
}
}
