% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_scraping.R
\name{ncaa_schools}
\alias{ncaa_schools}
\title{NCAA Schools}
\usage{
ncaa_schools()
}
\value{
a tibble with two columns, \code{school_name} and
\code{school_id}, both character-valued.
}
\description{
Gets a table of school names and IDs from the NCAA
}
\details{
based on \url{https://github.com/octonion/basketball-w/blob/master/ncaa/scrapers/ncaa_schools.rb}
}
\examples{
\dontrun{
school_list <- dfstools::ncaa_schools()
}
}
