% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_analytics.R
\name{nhl_goalie_archetypes}
\alias{nhl_goalie_archetypes}
\title{NHL Goalie Archetypal Analysis}
\usage{
nhl_goalie_archetypes(goalie_totals, num_archetypes = 2)
}
\arguments{
\item{goalie_totals}{a \code{goalie_totals} tibble returned by \code{nhl_player_season_totals}}

\item{num_archetypes}{number of archetypes to use (default 2)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item goalie_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object}
}
\description{
perform an "archetypal athletes" analysis
}
\examples{
\dontrun{
dfstools::msf_set_apikey("your MySportsFeeds API key")
player_totals <- dfstools::nhl_player_season_totals("2018-2019-regular")
the_archetypes <- dfstools::nhl_goalie_archetypes(player_totals$goalie_totals)
goalie_alphas <- the_archetypes[["player_alphas"]]
View(goalie_alphas)
}
}
