% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_analytics.R
\name{nhl_play_by_play}
\alias{nhl_play_by_play}
\title{NHL Play-by-play}
\usage{
nhl_play_by_play(season, game_slug)
}
\arguments{
\item{season}{the NHL season}

\item{game_slug}{a string of the form \code{yyyymmdd-AAA-HHH}, where \code{yyyymmdd}
is the game start date, \code{AAA} is the away team abbreviation, and \code{HHH} is the
home team abbreviation.}
}
\value{
a tibble of all the plays in the game
}
\description{
retrieve an NHL game play-by-play tibble
}
\examples{
\dontrun{
library(dplyr)
stanley_cup_games <- dfstools::msf_seasonal_games("nhl", "2019-playoff")
stanley_cup_final <- stanley_cup_games \%>\%
  dplyr::arrange(desc(slug)) \%>\%
  head(1)
stanley_cup_final_slug <- stanley_cup_final[["slug"]]
stanley_cup_final_pbp <-
  dfstools::nhl_play_by_play("2019-playoff", stanley_cup_final_slug)
}
}
