% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_analytics.R
\name{nhl_player_season_totals}
\alias{nhl_player_season_totals}
\title{NHL Player Season Totals}
\usage{
nhl_player_season_totals(season)
}
\arguments{
\item{season}{a valid MySportsFeeds v2.1 API season name}
}
\value{
a list of three items
\itemize{
\item goalie_totals a tibble of goalie season total box score statistics,
arranged by descending games won
\item skater_totals a tibble of skater season total box score statistics,
arranged by descending goals made
\item player_labels a tibble of labeling information for both goalies and
skaters
}
}
\description{
fetches the player season totals for base box score statistics
}
\examples{
\dontrun{
player_totals <- dfstools::nhl_player_season_totals("2018-2019-regular")
goalie_totals <- player_totals$goalie_totals
skater_totals <- player_totals$skater_totals
}
}
