% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_analytics.R
\name{nhl_skater_archetype_search}
\alias{nhl_skater_archetype_search}
\title{NHL Skater Archetype Search}
\usage{
nhl_skater_archetype_search(skater_totals, num_steps = 1:7, nrep = 32,
  verbose = FALSE)
}
\arguments{
\item{skater_totals}{a tibble returned by \code{nhl_player_season_totals}}

\item{num_steps}{number of steps to use (default 1:5)}

\item{nrep}{number of repetitions at each step (default 4)}

\item{verbose}{should the search be verbose? (default FALSE)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object - the \code{bestModel} with \code{num_steps}
archetypes
\item all of the models}
}
\description{
stepwise search of archetype counts
}
\examples{
\dontrun{
dfstools::msf_set_apikey("your MySportsFeeds API key")
player_totals <- dfstools::nhl_player_season_totals("2018-2019-regular")
goalie_totals <- player_totals$goalie_totals
skater_totals <- player_totals$skater_totals
skater_archetypes <- dfstools::nhl_skater_archetype_search(skater_totals)
screeplot(skater_archetypes$archetype_models)
skater_alphas <- skater_archetypes[["player_alphas"]]
View(skater_alphas)
}
}
