% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_analytics.R
\name{nhl_skater_archetypes}
\alias{nhl_skater_archetypes}
\title{NHL Skater Archetypal Analysis}
\usage{
nhl_skater_archetypes(skater_totals, num_archetypes = 3)
}
\arguments{
\item{skater_totals}{a \code{skaters_totals} tibble returned by \code{nhl_player_season_totals}}

\item{num_archetypes}{number of archetypes to use (default 3)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item skater_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object}
}
\description{
perform an "archetypal athletes" analysis
}
\examples{
\dontrun{
dfstools::msf_set_apikey("your MySportsFeeds API key")
player_totals <- dfstools::nhl_player_season_totals("2018-2019-regular")
the_archetypes <- dfstools::nhl_skater_archetypes(player_totals$skater_totals)
skater_alphas <- the_archetypes[["player_alphas"]]
View(skater_alphas)
}
}
