% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_entries.R
\name{parse_entries}
\alias{parse_entries}
\title{Parse dfs entry files}
\usage{
parse_entries(path = NULL, sport = NULL, platform = NULL)
}
\arguments{
\item{path}{character, path to salary file}

\item{sport}{character, which sports league? Supports nfl, nba on the way}

\item{platform}{charcter, which dfs platform? Supports draftkings, fanduel on the way}
}
\value{
data.frame of entry info
}
\description{
Parse dfs entry files
}
