% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_optimized_lineups.R
\name{parse_optimized_lineups}
\alias{parse_optimized_lineups}
\title{Parse dfs optimized lineup files}
\usage{
parse_optimized_lineups(path = NULL, sport = NULL, platform = NULL)
}
\arguments{
\item{path}{character, path to optimized lineups file}

\item{sport}{character, which sports league? Supports nfl, nba on the way}

\item{platform}{charcter, which dfs platform? Supports draftkings, fanduel on the way}
}
\value{
data.frame of optimized lineups
}
\description{
Parse dfs optimized lineup files
}
