% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_salaries.R
\name{parse_salaries}
\alias{parse_salaries}
\title{Parse dfs player salary files}
\usage{
parse_salaries(path = NULL, sport = NULL, platform = NULL)
}
\arguments{
\item{path}{character, path to salary file}

\item{sport}{character, which sports league? Supports nfl and nba for draftkings}

\item{platform}{charcter, which dfs platform? Supports draftkings}
}
\value{
data.frame of players with salary info
}
\description{
Parse dfs player salary files
}
