% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{player_box_score}
\alias{player_box_score}
\title{Make player box score data frame}
\usage{
player_box_score(spreadsheet, sheet_number = 1)
}
\arguments{
\item{spreadsheet}{a file with an BigDataBall player box score dataset}

\item{sheet_number}{the sheet number within the spreadsheet (default 1)}
}
\value{
a data frame of the input spreadsheet, augmented with columns for
double-doubles, triple doubles, "game scores" and fantasy points. Note that
since turnovers are bad, the turnover column is the negative of the numbe
of turnovers!
}
\description{
Creates a player box score data frame from an BigDataBall spreadsheet
}
\examples{
\dontrun{
playerbox <- player_box_score(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
}
}
