% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{playerboxscore}
\alias{playerboxscore}
\title{Make player box score data frame}
\usage{
playerboxscore(spreadsheet, sheet_number = 1)
}
\arguments{
\item{spreadsheet}{a file with an NBAStuffer player box score dataset}

\item{sheet_number}{the sheet number within the spreadsheet (default 1)}
}
\value{
a data frame of the input spreadsheet, augmented with columns for double-doubles, triple doubles and fantasy points
}
\description{
Creates a player box score data frame from an NBAStuffer spreadsheet
}
\examples{
\dontrun{
playerbox <- playerboxscore(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
}
}
