% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{project_season}
\alias{project_season}
\title{Project season}
\usage{
project_season(raw_season)
}
\arguments{
\item{raw_season}{a tibble returned from the Stattleship API via `get_season`}
}
\value{
a list with three items:
\itemize{
\item projections: a tibble with the projections for the rest of the season.
\item model: the model that mvglmmRank built from the closed games
\item model_input: model input data from the closed games}
}
\description{
Creates a tibble with projections for the rest of the games in a season
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_raw <-
  tidysportsfeeds::get_season(league = "nba")
nba_result <- project_season(nba_raw)
nba_projections <- nba_result$projections
readr::write_excel_csv(nba_projections, "nba_projections.csv")
}
}
