% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{project_upcoming_games}
\alias{project_upcoming_games}
\title{Project upcoming games for the current season}
\usage{
project_upcoming_games(games)
}
\arguments{
\item{games}{a `games` tibble returned from the Stattleship API via `get_games`}
}
\value{
a list with three items:
\itemize{
\item projections: a tibble with the projections for the rest of the season.
\item model: the model that mvglmmRank built from the closed games
\item model_input: model input data from the closed games}
}
\description{
Creates a tibble with projections for the rest of the games in a season
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
stattleshipR::set_token(token)
nba_games <- get_games(league = "nba")
nba_result <- project_upcoming_games(nba_games)
nba_projections <- nba_result$projections
readr::write_excel_csv(nba_projections, "nba_projections.csv")
}
}
