% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{rank_scores}
\alias{rank_scores}
\title{Rank predicted scores}
\usage{
rank_scores(aug_schedule)
}
\arguments{
\item{aug_schedule}{an augmented schedule from tidy_game_predict}
}
\value{
the teams ranked by ascending scores
}
\description{
rank the teams by predicted scores
}
