% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest_history.R
\name{read_draftkings_contest_history}
\alias{read_draftkings_contest_history}
\title{Read DraftKings Contest History}
\usage{
read_draftkings_contest_history(file)
}
\arguments{
\item{file}{CSV history file from DraftKings}
}
\value{
a tibble of the contest history
}
\description{
Read a DraftKings contest history CSV into a tibble
}
\examples{
\dontrun{
View(dfstools::read_draftkings_contest_history(
"~/Projects/draftkings/contest-standings/draftkings-contest-entry-history.csv"))
}
}
