% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdataball.R
\name{real_points}
\alias{real_points}
\title{Real-world points}
\usage{
real_points(boxscore)
}
\arguments{
\item{boxscore}{a player or game box score data frame}
}
\value{
a vector of points
}
\description{
Computes points from a box score
}
\examples{
\dontrun{
playerbox <- player_box_score(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
points <- real_points(playerbox)
}
}
