% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_utils.R
\name{select_nba_games_columns}
\alias{select_nba_games_columns}
\title{Select NBA Games Columns}
\usage{
select_nba_games_columns(games_object)
}
\arguments{
\item{games_object}{a `games` object returmed from msf_seasonal_games for the NBA!}
}
\value{
a `games` data frame with some columns removed
}
\description{
Extracts the relevant data from a MySportsFeeds NBA "games" object
}
\details{
The NBA `games` object that comes from the NBA has some columns with data issues. Some are all `NA`, and others are list columns that databases can't handle. So we remove them with `dplyr::select`.
}
