% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{set_msf_apikey}
\alias{set_msf_apikey}
\title{Set MySportsFeeds API key}
\usage{
set_msf_apikey(apikey)
}
\arguments{
\item{apikey}{the API key}
}
\description{
sets the MySportsFeeds API key into the keyring
}
