% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{sq_append_table}
\alias{sq_append_table}
\title{SQLite Append Table}
\usage{
sq_append_table(connection, table, dataframe)
}
\arguments{
\item{connection}{connection to the database}

\item{table}{the table name. It will be created if it doesn't
exist}

\item{dataframe}{the data frame to append}
}
\value{
a list of two items
\itemize{
\item table_create logical was the table created?
\item table_append integer the number of rows appended to the
table
}
}
\description{
creates a table if needed and appends a data
frame to it
}
