% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{sq_connect_dbfile}
\alias{sq_connect_dbfile}
\title{SQLite Connect Database File}
\usage{
sq_connect_dbfile(path)
}
\arguments{
\item{path}{full path to file. It will be created if it
doesn't exist}
}
\value{
a DBI connection object
}
\description{
creates an SQLite database file if needed and
returns its connection
}
\examples{
\dontrun{
connection <- dfstools::sq_connect_dbfile(
"~/DFS/nba/dfs_database.sqlite"
)
}
}
