% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_utils.R
\name{sq_create_databases}
\alias{sq_create_databases}
\title{SQLite Create Databases}
\usage{
sq_create_databases(directory, verbose)
}
\arguments{
\item{directory}{valid path to a directory where the database
files will be stored. It will be created if it doesn't exist.
The files will be overwritten if they exist and created if they
don't.}

\item{verbose}{print status information}
}
\value{
a list of the files in \code{directory} as given by
\code{list.files}.
}
\description{
Creates an SQLite database for each league /
season returned by \code{mfs_seasons()}.
}
