% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{stattleship_game_data}
\alias{stattleship_game_data}
\title{Stattleship collect game data}
\usage{
stattleship_game_data(stattleship_games)
}
\arguments{
\item{stattleship_games}{"games" tibble from Stattleship API}
}
\value{
a game data dataframe
}
\description{
creates an input game data file for mvglmmRank
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_stattleship_games <-
  tidysportsfeeds::get_stattleship_games(league = "nba")
nba_game_data <- tidysportsfeeds::stattleship_game_data(nba_stattleship_games)
}
}
