% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{stattleship_upcoming_games}
\alias{stattleship_upcoming_games}
\title{Stattleship upcoming games}
\usage{
stattleship_upcoming_games(stattleship_games)
}
\arguments{
\item{stattleship_games}{"games" tibble from Stattleship API}
}
\value{
a game data dataframe
}
\description{
creates a data frame with upcoming games
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_stattleship_games <-
  tidysportsfeeds::get_stattleship_games(league = "nba")
nba_game_data <- tidysportsfeeds::stattleship_upcoming_games(nba_stattleship_games)
}
}
