% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_for_optimization.R
\name{structure_for_optimization}
\alias{structure_for_optimization}
\title{Prepare a file for dfs optimization}
\usage{
structure_for_optimization(
  salaries = NULL,
  projections = NULL,
  platform = NULL,
  sport = NULL
)
}
\arguments{
\item{salaries}{data.frame of salary data}

\item{projections}{data.frae with projections}

\item{platform}{character, dfs platform}

\item{sport}{character, sport}
}
\value{
data.frame to be used as input for optimization
}
\description{
Prepare a file for dfs optimization
}
