% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_optimized_lineups.R
\name{structure_optimized_entries}
\alias{structure_optimized_entries}
\title{Insert optimized lineups into entries}
\usage{
structure_optimized_entries(
  entries = NULL,
  optimized_lineups = NULL,
  platform = NULL,
  sport = NULL
)
}
\arguments{
\item{entries}{data.frame of entries}

\item{optimized_lineups}{data.frame of optimized lineups}

\item{platform}{character platform}

\item{sport}{character sport}
}
\value{
data.frame of entries with optimized lineups
}
\description{
Insert optimized lineups into entries
}
