% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Ternary Plot}
\usage{
ternary_plot(player_alphas, plot_title)
}
\arguments{
\item{player_alphas}{a data frame with player archetype values. The player
name must be in column \code{player_name}.
There must be exactly three archetypes, in order \code{Rim Protection},
\code{Floor Spacing} and \code{Bench}.}

\item{plot_title}{the plot title}
}
\value{
a \code{ggplot} object
}
\description{
a visualization of a set of players using \code{ggtern}
}
