% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wnba_2020_first_nine_games_player_totals}
\alias{wnba_2020_first_nine_games_player_totals}
\title{WNBA 2020 Player Totals After Nine Games}
\format{
A data frame with 144 rows and 27 variables:
\describe{
\item{player}{player name}
\item{tm}{team code name}
\item{pos}{position(s)}
\item{g}{games played}
\item{gs}{games started}
\item{mp}{minutes played}
\item{fg}{field goals made}
\item{fga}{field goals attempted}
\item{fg_percent}{field goal percentage}
\item{x3p}{three-point field goals made}
\item{x3pa}{three-point field goals attempted}
\item{x3p_percent}{three-point percentage}
\item{x2p}{two-point field goals made}
\item{x2pa}{two-point field goals attempted}
\item{x2p_percent}{two-point field goal percentage}
\item{e_fg_percent}{effective field goal percentage}
\item{ft}{free throws made}
\item{fta}{free throws attempted}
\item{ft_percent}{free throw percentage}
\item{orb}{offensive rebounds}
\item{trb}{total rebounds - offensive plus defensive}
\item{ast}{assists}
\item{stl}{steals}
\item{blk}{blocks}
\item{tov}{turnovers}
\item{pf}{personal fouls}
\item{pts}{points}
}
}
\source{
\url{https://www.basketball-reference.com/wnba/years/2020_totals.html}
}
\usage{
wnba_2020_first_nine_games_player_totals
}
\description{
A dataset containing the player total statistics for the first nine games of
the WNBA 2020 season. This is used as a test / sample dataset for the
archetypal analysis functions.
}
\keyword{datasets}
