% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_bbref.R
\name{wnba_2020_rank_prep_bbref}
\alias{wnba_2020_rank_prep_bbref}
\title{WNBA 2020 rank prep from basketball-reference.com}
\usage{
wnba_2020_rank_prep_bbref()
}
\value{
a list with two items
\itemize{
\item game.data a \code{game.data} table
\item schedule a \code{schedule} table
}
}
\description{
fetches the 2020 WNBA schedule from basketball-reference.com and
converts it to tables usable by \code{dfstools::mvglmmRank_model} and
\code{dfstools::game_predict}
}
