% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_com.R
\name{wnba_2020_rank_prep_wnba}
\alias{wnba_2020_rank_prep_wnba}
\title{WNBA 2020 rank prep from wnba.com}
\usage{
wnba_2020_rank_prep_wnba()
}
\value{
a list with three items
\itemize{
\item game.data a \code{game.data} table
\item games_played a table with the number of games each team has played
\item schedule a \code{schedule} table
}
}
\description{
fetches the 2020 WNBA schedule from stats.wnba.com and
converts it to tables usable by \code{dfstools::mvglmmRank_model} and
\code{dfstools::game_predict}
}
