% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_2020_totals_bbref}
\alias{wnba_2020_totals_bbref}
\title{WNBA 2020 player totals from Basketball Reference}
\usage{
wnba_2020_totals_bbref()
}
\value{
a list of two items
\itemize{
\item player_totals a tibble of player season total box score statistics,
arranged by descending points scored
\item player_labels a tibble of labeling information for players
}
}
\description{
fetches the 2020 WNBA player totals from
basketball-reference.com and prepares the raw data for archetypal
analysis.
}
\examples{
\dontrun{
wnba_totals <- dfstools::wnba_2020_totals_bbref()
player_totals <- wnba_totals$player_totals
player_labels <- wnba_totals$player_labels
the_archetypes <- dfstools::wnba_archetypes(player_totals)
player_alphas <- the_archetypes[["player_alphas"]]
View(player_alphas)
}
}
