% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_archetype_search}
\alias{wnba_archetype_search}
\title{WNBA Archetype Search}
\usage{
wnba_archetype_search(
  player_totals,
  num_steps = 1:10,
  nrep = 64,
  verbose = TRUE
)
}
\arguments{
\item{player_totals}{a tibble returned by \code{nba_player_season_totals}}

\item{num_steps}{number of steps to use (default 1:10)}

\item{nrep}{number of repetitions at each step (default 64)}

\item{verbose}{should the search be verbose? (default TRUE)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object - the \code{bestModel} with \code{num_steps}
archetypes
\item all of the models}
}
\description{
stepwise search of archetype counts
}
\examples{
\dontrun{
wnba_totals <- dfstools::wnba_season_totals_bbref(2020)
player_totals <- wnba_totals$player_totals
player_labels <- wnba_totals$player_labels
the_archetypes <- dfstools::wnba_archetype_search(player_totals)
screeplot(the_archetypes$archetype_models)
}
}
