% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_analytics.R
\name{wnba_archetypes}
\alias{wnba_archetypes}
\title{WNBA Archetypal Analysis}
\usage{
wnba_archetypes(player_totals, num_archetypes = 3)
}
\arguments{
\item{player_totals}{a tibble returned by \code{nba_player_season_totals}}

\item{num_archetypes}{number of archetypes to use (default 3)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object}
}
\description{
perform an "archetypal athletes" analysis
}
\examples{
\dontrun{
wnba_totals <- dfstools::wnba_season_totals_bbref(2020)
player_totals <- wnba_totals$player_totals
player_labels <- wnba_totals$player_labels
the_archetypes <- dfstools::wnba_archetypes(player_totals)
player_alphas <- the_archetypes[["player_alphas"]]
archetype_parameters <- the_archetypes[["archetype_parameters"]]
View(player_alphas)
View(archetype_parameters)
}
}
