% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_bdb_dfs}
\alias{wnba_bdb_dfs}
\title{WNBA BigDataBall DFS}
\usage{
wnba_bdb_dfs(bdb_excel_file)
}
\arguments{
\item{bdb_excel_file}{a BigDataBall WNBA DFS file}
}
\value{
a tibble with the DFS data
}
\description{
fetches a BigDataBall WNBA player box score
}
\examples{
\dontrun{
wnba_dfs <- dfstools::wnba_bdb_dfs(
  "~/Dropbox/20-wnba-dfs/08-07-2020-wnba-season-dfs-feed.xlsx"
)
View(wnba_dfs)
}
}
