% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_bdb.R
\name{wnba_bdb_player_box_score}
\alias{wnba_bdb_player_box_score}
\title{WNBA BigDataBall Player Box Score}
\usage{
wnba_bdb_player_box_score(bdb_excel_file)
}
\arguments{
\item{bdb_excel_file}{a BigDataBall WNBA player box score file}
}
\value{
a tibble with the player box score
}
\description{
fetches a BigDataBall WNBA player box score
}
\examples{
\dontrun{
wnba_player_box_score <- dfstools::wnba_bdb_player_box_score(
  "~/Dropbox/20-wnba-player/08-07-2020-wnba-season-player-feed.xlsx"
)
View(wnba_player_box_score)
}
}
