% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_bdb_team_box_score}
\alias{wnba_bdb_team_box_score}
\title{WNBA BigDataBall Team Box Score}
\usage{
wnba_bdb_team_box_score(bdb_excel_file)
}
\arguments{
\item{bdb_excel_file}{a BigDataBall WNBA team box score file}
}
\value{
a tibble with the team box score
}
\description{
fetches a BigDataBall WNBA team box score
}
\examples{
\dontrun{
wnba_team_box_score <- dfstools::wnba_bdb_team_box_score(
  "~/Dropbox/20-wnba-team/08-09-2020-wnba-season-team-feed.xlsx"
)
View(wnba_team_box_score)
}
}
