% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_make_game_data}
\alias{wnba_make_game_data}
\title{Make WNBA game.data}
\usage{
wnba_make_game_data(wnba_schedule)
}
\arguments{
\item{wnba_schedule}{a tibble returned from \code{wnba_2020_schedule}}
}
\value{
a \code{game.data} table
}
\description{
Builds a \code{mvglmmRank} "game.data" table from a
\code{wnba_2020_schedule} result
}
\examples{
\dontrun{
wnba_schedule <- dfstools::wnba_2020_schedule()
wnba_game_data <- dfstools::wnba_make_game_data(wnba_schedule)
View(wnba_game_data)
}
}
