% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_make_game_predict_schedule}
\alias{wnba_make_game_predict_schedule}
\title{Make WNBA \code{game_predict} schedule}
\usage{
wnba_make_game_predict_schedule(wnba_schedule)
}
\arguments{
\item{wnba_schedule}{a tibble returned from \code{wnba_2020_schedule}}
}
\value{
a \code{schedule} table
}
\description{
Builds a \code{game_predict} \code{schedule} table from a
\code{wnba_2020_schedule} result
}
\examples{
\dontrun{
wnba_schedule <- dfstools::wnba_2020_schedule()
wnba_game_data <- dfstools::wnba_make_game_data(wnba_schedule)
wnba_game_predict_schedule <-
  dfstools::wnba_make_game_predict_schedule(wnba_schedule)
View(wnba_game_data)
View(wnba_game_predict_schedule)
}
}
