% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_rest_days}
\alias{wnba_rest_days}
\title{Rest days}
\usage{
wnba_rest_days(schedule)
}
\arguments{
\item{schedule}{a tibble returned by \code{wnba_2020_schedule_bbref}}
}
\value{
a tibble with three columns
\itemize{
\item \code{team} character team name,
\item \code{date} POSIXct the game date,
\item \code{rest_days} numeric number of rest days
}
}
\description{
Compute rest days from the schedule
}
\examples{
\dontrun{
wnba_schedule <- dfstools::wnba_2020_schedule_bbref()
View(wnba_schedule)
rest_days <- dfstools::wnba_rest_days(wnba_schedule)
View(rest_days)
}
}
