% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_com.R
\name{wnba_season_totals_wnba}
\alias{wnba_season_totals_wnba}
\title{WNBA season player totals from wnba.com}
\usage{
wnba_season_totals_wnba(season)
}
\arguments{
\item{season}{season to fetch (1997 - 2020)}
}
\value{
a list of two items
\itemize{
\item player_totals a tibble of player season total box score statistics,
arranged by descending minutes played
\item player_labels a tibble of labeling information for players
}
}
\description{
fetches the WNBA player totals for a given season from
wnba.com and prepares the raw data for archetypal
analysis.
}
\examples{
\dontrun{
wnba_2020_totals_wnba <- dfstools::wnba_season_totals_wnba(2020)
archetype_search_result <-
  dfstools::archetype_search(wnba_2020_totals_wnba$player_totals)
screeplot(archetype_search_result$archetype_models)
the_archetypes <-
  dfstools::compute_archetypes(wnba_2020_totals_wnba$player_totals)
View(the_archetypes$player_alphas)
View(the_archetypes$archetype_parameters)
}
}
