% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{yahoo_points}
\alias{yahoo_points}
\title{Yahoo fantasy points}
\usage{
yahoo_points(boxscore)
}
\arguments{
\item{boxscore}{a player box score data frame}
}
\value{
a vector of Yahoo fantasy points
}
\description{
Computes Yahoo fantasy points from a box score
}
\examples{
\dontrun{
playerbox <- playerboxscore(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
yh_points <- yahoo_points(playerbox)
}
}
