% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{archetype_search}
\alias{archetype_search}
\title{Archetype Search}
\usage{
archetype_search(player_totals, num_steps = 1:10, nrep = 64, verbose = FALSE)
}
\arguments{
\item{player_totals}{a tibble of the values to use for archetypal analysis.
The first column must be the player name. The archetypes will be sorted on
the second columns. Usually, this will be a measure of the player's total
activity for the season, like minutes played or plate appearances / batters
faced.

If the column \code{team_games_played} exists, the other numeric values will
be divided by it and the column will be deleted before doing the search.
\code{team_games_played} must contain the number of games the player's team
has played as of the player total dataset acquisition.}

\item{num_steps}{number of steps to use (default 1:10)}

\item{nrep}{number of repetitions at each step (default 64)}

\item{verbose}{should the search be verbose? (default FALSE)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object - the \code{bestModel} with \code{max(num_steps)}
archetypes
\item all of the models}
}
\description{
stepwise search of archetype counts
}
