% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{archetype_search}
\alias{archetype_search}
\title{Archetype Search}
\usage{
archetype_search(pbs)
}
\arguments{
\item{pbs}{player box score}
}
\value{
an "as" object from archetypes::stepArchetypes
}
\description{
search for the best archetypal analysis of real box score values
}
