% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{build_model}
\alias{build_model}
\title{Build model}
\usage{
build_model(game_data, method = "PB1", first.order = TRUE)
}
\arguments{
\item{game_data}{a game.data input file}

\item{method}{method to be passed to mvglmmRank - default is "N"}

\item{first.order}{flag to be passed to mvglmmRank - default is TRUE}
}
\value{
an mvglmmRank model object
}
\description{
runs mvglmmRank::mvglmmRank against a game box score and points function
}
