% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{compute_archetypes}
\alias{compute_archetypes}
\title{Archetypal Analysis}
\usage{
compute_archetypes(player_totals, num_archetypes = 3)
}
\arguments{
\item{player_totals}{a tibble of the values to use for archetypal analysis.
The first column must be the player name. The archetypes will be sorted on
the second columns. Usually, this will be a measure of the player's total
activity for the season, like minutes played or plate appearances / batters
faced.}

\item{num_archetypes}{number of archetypes to use (default 3)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object}
}
\description{
perform an "archetypal athletes" analysis
}
