% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{game_data}
\alias{game_data}
\title{Collect game data}
\usage{
game_data(gbs, pfunc, end_date)
}
\arguments{
\item{gbs}{game box score data frame}

\item{pfunc}{a function mapping a box score data frame to a points vector, usually either real_points, draftkings_points, fanduel_points or yahoo_points}

\item{end_date}{the latest date to use as input}
}
\value{
a game data dataframe
}
\description{
creates an input game data file for mvglmmRank
}
