% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{game_predict}
\alias{game_predict}
\title{Game predict}
\usage{
game_predict(model, season)
}
\arguments{
\item{model}{a model from build_model}

\item{season}{a tibble returned from the Stattleship API via `get_games`}
}
\value{
the upcoming games from the season augmented with prediction columns
}
\description{
Creates a data frame with game predictions
}
