% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbastuffer.R
\name{gameboxscore}
\alias{gameboxscore}
\title{Make game box score data frame from player box score}
\usage{
gameboxscore(pbs)
}
\arguments{
\item{pbs}{a player box score data frame}
}
\value{
a data frame of game box scores
}
\description{
Creates a game box score data frame from a player box score data frame
}
\examples{
\dontrun{
playerbox <- playerboxscore(
  "~/Dropbox/16-17-player-data/season-player-feed-02-03-2017.xlsx")
gamebox <- gameboxscore(playerbox)
}
}
