% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_game_logs}
\alias{get_game_logs}
\title{Get current season game logs (player box scores) from Stattleship API}
\usage{
get_game_logs(league, team_slug)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}

\item{team_slug}{the Stattleship team slug for the team}
}
\value{
a tibble with the game logs
}
\description{
Gets a game_logs object from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
trailblazers_logs <-
  tidysportsfeeds::get_game_logs(league = "nba", team_slug = "nba-por")
}
}
