% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_game_logs_to_date}
\alias{get_game_logs_to_date}
\title{Get current season game logs to date from Stattleship API}
\usage{
get_game_logs_to_date(league)
}
\arguments{
\item{league}{("nba", "nhl" -- "nfl", "mlb" TBD!)}
}
\value{
a list with three items:
\itemize{
\item players the players tibble returned by the API
\item games the games tibble returned by the API
\item game_logs the game logs (player box scores) as delivered by the API}
}
\description{
Gets game logs to date for all teams from the stattleship.com API. Note: this is a lengthy process; the API delivers only one team at a time. If it crashes, you'll need to re-run it.
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
stattleshipR::set_token(token)
nba_data <- get_game_logs_to_date(league = "nba")
nhl_data <- get_game_logs_to_date(league = "nhl")
}
}
