% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_msf_api}
\alias{get_msf_api}
\title{GET from MySportsFeeds API}
\usage{
get_msf_api(url, verbose = FALSE)
}
\arguments{
\item{url}{the URL to GET}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
  \item status_code the HTTP status code (200 for success)
  \item response if status_code == 200, the JSON object;
    otherwise, the raw text.
}
}
\description{
GETs data from the MySportsFeeds 2.0 API
}
\examples{
\dontrun{
response <- dfstools::get_msf_api(
"https://api.mysportsfeeds.com/v2.0/pull/nba/2018-playoff/games.json"
)}
}
