% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_mysportsfeeds_dfs}
\alias{get_mysportsfeeds_dfs}
\title{Get MySportsFeed DFS data}
\usage{
get_mysportsfeeds_dfs(league, season)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}

\item{season}{Look up season code in API docs https://www.mysportsfeeds.com/data-feeds/api-docs}
}
\value{
a list of tibbles with the DFS data for the league and season. There is one tibble for each DFS site, currently "DraftKings" and "FanDuel".
}
\description{
Gets DFS data object from the MySportsFeeds.com API
}
\examples{
\dontrun{
username <- "your_user_name"
password <- "your_password"
library(dfstools)
mysportsfeedsR::authenticate_v1_0(username, password)
nba_dfs_2017_2018 <-
  get_mysportsfeeds_dfs(league = "nba", season = "2017-2018-regular")
}
}
