% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_season}
\alias{get_season}
\title{Get Current Season tables from Stattleship API}
\usage{
get_season(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a list of three tibbles:
\itemize{
\item teams: the teams in the league
\item players: the players in the league
\item games: games in the season.
All games - completed, in-progress and upcaming - are listed.}
}
\description{
Gets a schedule 'gameentry' object from the stattleship.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_season <-
  tidysportsfeeds::get_season(league = "nba")
nhl_season <-
  tidysportsfeeds::get_season(league = "nhl")
nfl_season <-
  tidysportsfeeds::get_season(league = "nfl")
}
}
