% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_season}
\alias{get_season}
\title{Get current season tables from Stattleship API}
\usage{
get_season(league)
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}
}
\value{
a list of three tibbles:
\itemize{
\item teams: the teams in the league
\item players: the players in the league
\item games: games in the season.
All games - closed, in-progress and upcaming - are listed.}
}
\description{
Gets a baseline list of tables for a season from the
Stattleship API. Usually, one runs this once at the beginning of the season
and updates the tables on a daily or weekly basis.
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_season <-
  get_season(league = "nba")
nhl_season <-
  get_season(league = "nhl")
nfl_season <-
  get_season(league = "nfl")
}
}
