% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{get_stattleship_games}
\alias{get_stattleship_games}
\title{Get Stattleship Games}
\usage{
get_stattleship_games(league = "nba", sport = "basketball")
}
\arguments{
\item{league}{("nba", "nhl", "nfl" or "mlb")}

\item{sport}{e.g., "basketball"}
}
\value{
a tibble with the games for the season. The whole schedule is given,
with games as yet unplayed having scores of zero.
}
\description{
Gets a schedule 'gameentry' object from the MySportsFeeds.com API
}
\examples{
\dontrun{
token <- "yourtoken"
library(tidysportsfeeds)
library(stattleshipR)
stattleshipR::set_token(token)
nba_stattleship_games <-
  tidysportsfeeds::get_stattleship_games(league = "nba", sport = "basketball")
nhl_stattleship_games <-
  tidysportsfeeds::get_stattleship_games(league = "nhl", sport = "hockey")
nfl_stattleship_games <-
  tidysportsfeeds::get_stattleship_games(league = "nfl", sport = "football")
}
}
