% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{make_game_data}
\alias{make_game_data}
\title{Make game.data}
\usage{
make_game_data(seasonal_games)
}
\arguments{
\item{seasonal_games}{a tibble returned from \code{msf_seasonal_games}}
}
\value{
a "game data" table
}
\description{
Builds a \code{mvglmmRank} "game.data" table from an
\code{msf_seasonal_games}result
}
\examples{
\dontrun{
nba_seasonal_games <- dfstools::msf_seasonal_games("nba", "2018-2019-regular")
nhl_seasonal_games <- dfstools::msf_seasonal_games("nhl", "2019-2020-regular")
nba.game.data <- dfstools::make_game_data(nba_seasonal_games)
nhl.game.data <- dfstools::make_game_data(nhl_seasonal_games)
}
}
