% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_past_nba_games}
\alias{msf_past_nba_games}
\title{MySportsFeeds Past NBA Games}
\usage{
msf_past_nba_games(seasons)
}
\arguments{
\item{seasons}{vector of season codes}
}
\value{
a tibble of completed NBA games
}
\description{
Returns a data frame of past NBA games from MySportsFeeds
}
\details{
`msf_past_nba_games` adds two columns at the right of the tibble:
  `season`, the source season of the row, and `startEastern`, the game
  start time in the Eastern USA timezone ("EST5EDT").
}
\examples{
\dontrun{
apikey <- "your_API key"
library(dfstools)
library(mysportsfeedsR)
authenticate_v2_x(apikey)
seasons <- c(
  "2018-playoff",
  "2017-2018-regular",
  "2017-playoff",
  "2016-2017-regular",
  "2016-playoff",
  "2015-2016-regular")
nba_games <- msf_past_nba_games(seasons)
}
}
