% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_games}
\alias{msf_seasonal_games}
\title{MySportsFeeds Seasonal Games}
\usage{
msf_seasonal_games(league, season, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{verbose}{print status info}
}
\value{
a tibble of games
}
\description{
Returns a data frame of games from MySportsFeeds version 2.1 API
}
\details{
\code{msf_seasonal_games} adds four columns at the right of the tibble:
\itemize{
\item league the source league of the data
\item season the source season of the data, and
\item date the game date (started) in the Eastern USA
timezone ("EST5EDT").
\item slug the game slug (date-away_team-home_team)
}

The MySportsFeeds API returns a list column with the scores in each inning
(MLB), quarter (NFL and NBA) or period (NHL). This column is unpacked with
\code{unnest} row-wise, so there is a row for each inning / quarter / period in
the game, including extra innings / overtime.

The returned tibble will be sorted in order of scheduled start time.
}
\examples{
\dontrun{
nba_games <- dfstools::msf_seasonal_games(
  season = "2017-2018-regular", league = "nba"
)}
}
