% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_player_gamelogs}
\alias{msf_seasonal_player_gamelogs}
\title{MySportsFeeds Seasonal Player Gamelogs}
\usage{
msf_seasonal_player_gamelogs(season, league, team, verbose = TRUE,
  tries = 5, sleep_seconds = 10)
}
\arguments{
\item{season}{the season to fetch}

\item{league}{the league to fetch}

\item{team}{the team to fetch}

\item{verbose}{print status info}

\item{tries}{number of times to retry on failure}

\item{sleep_seconds}{seconds to sleep after failure}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success, -200 for HTTP success but no DFS data))
\item response if status_code == 200, a tibble of DFS data; otherwise, the raw text.
}
}
\description{
Gets gamelogs object from from MySportsFeeds version 2.0 API
}
\examples{
\dontrun{
nba_gamelogs <- dfstools::msf_seasonal_player_gamelogs(
  season = "2018-playoff",
  league = "nba",
  team = "GSW"
)
}
}
