% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_player_stats_totals}
\alias{msf_seasonal_player_stats_totals}
\title{MySportsFeeds Seasonal Player Stats Totals}
\usage{
msf_seasonal_player_stats_totals(league, season, verbose = TRUE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{verbose}{print status info}
}
\value{
a \code{player_stats_totals} object
}
\description{
Gets player gamelogs object from from
MySportsFeeds version 2.1 API
}
\examples{
\dontrun{
nba_player_totals <- dfstools::msf_seasonal_player_stats_totals(
  season = "2018-2019-regular", league = "nba", verbose = TRUE
)}
}
