% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_players}
\alias{msf_seasonal_players}
\title{MySportsFeeds Seasonal Players}
\usage{
msf_seasonal_players(league, season, verbose = FALSE)
}
\arguments{
\item{league}{the league to fetch}

\item{season}{the season to fetch}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success)
\item response if status_code == 200, a tibble of players;
otherwise, the raw text.
}
}
\description{
Returns a data frame of players from
MySportsFeeds version 2.0 API
}
\examples{
\dontrun{
nba_players <- dfstools::msf_seasonal_players(
season = "2017-2018-regular", league = "nba"
)}
}
