% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_utils.R
\name{msf_seasonal_team_dfs}
\alias{msf_seasonal_team_dfs}
\title{MySportsFeeds Seasonal Team DFS}
\usage{
msf_seasonal_team_dfs(season, league, team, verbose = FALSE)
}
\arguments{
\item{season}{the season to fetch}

\item{league}{the league to fetch}

\item{team}{the team to fetch}

\item{verbose}{print status info}
}
\value{
a list of two items
\itemize{
\item status_code the HTTP status code (200 for success,
-200 for HTTP success but no DFS data)
\item response if status_code == 200, a tibble of DFS data;
otherwise, the raw response text.
}
}
\description{
Gets DFS data object from from MySportsFeeds version 2.0 API
}
\examples{
\dontrun{
nba_dfs <- dfstools::msf_seasonal_team_dfs(
season = "2018-playoff", league = "nba", team = "GSW"
)}
}
