% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_analytics.R
\name{nba_archetypes}
\alias{nba_archetypes}
\title{NBA Archetypal Analysis}
\usage{
nba_archetypes(player_totals, num_archetypes = 3)
}
\arguments{
\item{player_totals}{a tibble returned by \code{nba_player_season_totals}}

\item{num_archetypes}{number of archetypes to use (default 3)}
}
\value{
a list of
\itemize{
\item archetype_parameters the parameters that define each archetype
\item player_alphas the players tagged with their loadings on each archetype
\item archetype_model the model object}
}
\description{
perform an "archetypal athletes" analysis
}
\examples{
\dontrun{
dfstools::msf_set_apikey("your MySportsFeeds API key")
player_totals <- dfstools::nba_player_season_totals("2018-2019-regular")
the_archetypes <- dfstools::nba_archetypes(player_totals$player_totals)
player_alphas <- the_archetypes[["player_alphas"]]
View(player_alphas)
}
}
