% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_utils.R
\name{project_season}
\alias{project_season}
\title{Project season}
\usage{
project_season(model, season)
}
\arguments{
\item{model}{a model from build_model}

\item{season}{a tibble returned from the Stattleship API via `get_seaon`}
}
\value{
the upcoming games from the season augmented with projection columns:
\itemize{
\item started_at: the scheduled game start time
\item away: The away team name
\item home: the home team name
\item method: The method used to build the `mvglmmRank` model; the default is "PB1".
\item away_score_p: the projected score for the away team
\item home_score_p: the projected score for the home team
\item total_p: the projected total score
\item home_mov_p: the projected home margin of victory (home score - away score)
\item home_prob_w: the projected probability that the home team wins
\item away_prob_w: the projected probability that the away team wins
\item entropy: the projected Shannon entropy for the game}
}
\description{
Creates a tibble with projections for the rest of the games in a season
}
