% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_models.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Ternary Plot}
\usage{
ternary_plot(player_table, plot_title, top_label, left_label)
}
\arguments{
\item{player_table}{a data frame with player archetype values}

\item{plot_title}{the plot title}

\item{top_label}{the label for the top apex (backcourt)}

\item{left_label}{the label for the left apex (frontcourt)}
}
\value{
a `ggplot` object
}
\description{
a visualization of a set of players using `ggtern`
}
