% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_2020_schedule_bbref}
\alias{wnba_2020_schedule_bbref}
\title{Basketball Reference WNBA Schedule}
\usage{
wnba_2020_schedule_bbref()
}
\value{
a tibble with three columns
\itemize{
\item \code{date} POSIXct the game date,
\item \code{away} character the "away" team
\item \code{home} character the "home" team
}
}
\description{
fetches the 2020 WNBA schedule from Basketball-reference.com and
converts it to a format usable by dfstools::mvglmmRank_model
}
\examples{
\dontrun{
wnba_schedule <- dfstools::wnba_2020_schedule_bbref()
View(wnba_schedule)
}
}
