% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_analytics.R
\name{wnba_ranking}
\alias{wnba_ranking}
\title{WNBA Ranking}
\usage{
wnba_ranking(rank_prep, method = "NB")
}
\arguments{
\item{rank_prep}{the output of \code{dfstools::wnba_rank_prep_wnba} or
\code{dfstools::wnba_rank_prep_bbref}}

\item{method}{The method to use in \code{mvglmmRank}. Default is "NB".}
}
\value{
a list of five items
\itemize{
\item rank_model the result of mvglmmRank_model
\item offensive ratings
\item defensive ratings
\item forecast the result of game_predict
\item daily_entropy the projected entropy for upcoming dates
}
}
\description{
Runs \code{dfstools::mvglmmRank} and \code{dfstools::game_predict} and
returns a list of the results
}
\examples{
\dontrun{
rank_prep_wnba <- dfstools::wnba_2020_rank_prep_wnba()
ranking_wnba <- dfstools::wnba_ranking(rank_prep_wnba)
rank_prep_bbref <- dfstools::wnba_2020_rank_prep_bbref()
ranking_bbref <- dfstools::wnba_ranking(rank_prep_bbref)
}
}
