% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba.R
\name{wnba_season_totals_wnba}
\alias{wnba_season_totals_wnba}
\title{WNBA season player totals from wnba.com}
\usage{
wnba_season_totals_wnba(season)
}
\arguments{
\item{season}{season to fetch (1997 - 2020)}
}
\value{
a list of two items
\itemize{
\item player_totals a tibble of player season total box score statistics,
arranged by descending points scored
\item player_labels a tibble of labeling information for players
}
}
\description{
fetches the WNBA player totals for a given season from
wnba.com and prepares the raw data for archetypal
analysis.
}
\examples{
\dontrun{
wnba_totals <- dfstools::wnba_season_totals_wnba(2020)
player_totals <- wnba_totals$player_totals
player_labels <- wnba_totals$player_labels
the_archetypes <- dfstools::wnba_archetypes(player_totals)
player_alphas <- the_archetypes[["player_alphas"]]
archetype_parameters <- the_archetypes[["archetype_parameters"]]
View(player_alphas)
View(archetype_parameters)
}
}
